requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
einops>=0.7.0
matplotlib>=3.7.0
lpips>=0.1.4
scikit-image>=0.21.0
pytest>=7.4.0
